requirements.txt
aiosqlite
fastapi
httpx
ipython
strawberry-graphql[fastapi]
typer[all]
uvicorn[standard]
SQLAlchemy